-Split Octo file.
